Initial Configuration with documentation
This is the initial configuration of the SB dictionary. The IngestLDD file should work, the documentation has been thrown together from the descriptions in the IngestLDD, primarily, and a rudimentary VALID test added to the test/ folder. The supplied stub FAIL and VALID tests have been renamed to, I hope, prevent them being run and messing up the build.