Update julian dates rule to workaround Saxon bug
For some reason, the Saxon reader does not like using the `.` for querying all the XML elements. Instead, it seems to prefer `*`.

For that reason, this rule was failing:

```
sb:SB_Metadata//.[ends-with(name(),'_JD')]
```

But now this works just fine:

```
sb:SB_Metadata//*[ends-with(name(),'_JD')]
```

Not sure why, but it works.